+++ keta.js
@@ -492,7 +492,9 @@ angular.module('keta.directives.AppBar',
 				var defaultLinks = {
 					ALL_APPS: null,
 					ALL_ENERGY_MANAGERS: null,
-					APP_ROOT: CommonUtils.addUrlParameter('/', 'userId', accessToken.user_id),
+					APP_ROOT:
+						accessToken !== null && angular.isDefined(accessToken.user_id) ?
+							CommonUtils.addUrlParameter('/', 'userId', accessToken.user_id) : '/',
 					USER_PROFILE: null,
 					USER_LOGOUT: null
 				};
@@ -4170,13 +4172,32 @@ angular.module('keta.directives.TimeRangeSelector', [
 					var yearsChunked = [];
 
 					var now =
-						moment(scope.displayValue)
+						moment(new Date())
 							.month(0).date(1).hour(0).minute(0).second(0).millisecond(0)
 							.toDate();
 
 					var currentYear = parseInt($filter('date')(now, 'yyyy'), 10);
 					var yearStart = currentYear - scope.yearsBefore;
 					var yearEnd = currentYear + scope.yearsAfter;
+
+					var viewYear = parseInt($filter('date')(scope.displayValue, 'yyyy'), 10);
+					var range = scope.yearsBefore + scope.yearsAfter + 1;
+
+					// shift to the past until view year is within range
+					if (viewYear < yearStart) {
+						while (viewYear < yearStart) {
+							yearStart -= range;
+							yearEnd -= range;
+						}
+					}
+
+					// shift to the future until view year is within range
+					if (viewYear > yearEnd) {
+						while (viewYear > yearEnd) {
+							yearStart += range;
+							yearEnd += range;
+						}
+					}
 
 					for (var i = yearStart; i <= yearEnd; i++) {
 						now.setFullYear(i);
@@ -4514,6 +4535,11 @@ angular.module('keta.directives.TimeRangeSelector', [
 							scope.model.from, scope.model.to,
 							scope.minRangeLength, scope.maxRangeLength
 						);
+
+						// update display value
+						scope.displayValue =
+							lastSelected === LAST_SELECTED_FROM ?
+								angular.copy(scope.model.from) : angular.copy(scope.model.to);
 
 					}
 				};